requirements.txt
pandas
openpyxl
scapy
pyinstaller
python-nmap
